requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==45.0.5
cssselect==1.3.0
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.6
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
futures==3.0.5
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.33.2
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
langchain-core==0.3.68
langchain-google-genai==2.0.8
langchain-openai==0.2.14
langsmith==0.4.4
litellm==1.60.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
nltk==3.9.1
numpy==2.2.6
openai==1.60.2
orjson==3.10.18
packaging==24.2
pillow==10.4.0
playwright==1.53.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.41.3
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.13.0
lxml
futures
langchain-openai
google-generativeai
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
fuzzywuzzy
PyMuPDF
python-docx
python-multipart
crawl4ai
playwright
colorama
anthropic
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
